% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{bouts}
\alias{bouts}
\title{Identify Bouts of Physical Activity Behaviors}
\usage{
bouts(
  mets,
  wake = NULL,
  bout_length = 10,
  thresh_lower = 1.25,
  thresh_upper = 6,
  tol = 0,
  tol_lower = 1.25,
  tol_upper = 3,
  epoch = 1
)
}
\arguments{
\item{mets}{Numeric vector of METs data.}

\item{wake}{Optional numeric vector indicating wear time (1 for wear, 0 for non-wear). If NULL, all data is considered valid wear time.}

\item{bout_length}{Numeric value indicating the minimum length of a bout in minutes. Default is 10 minutes.}

\item{thresh_lower}{Numeric value indicating the lower threshold for valid METs. Default is 1.25.}

\item{thresh_upper}{Numeric value indicating the upper threshold for valid METs. Default is 6.}

\item{tol}{Numeric value indicating the tolerance for non-valid METs within a bout. Default is 0.}

\item{tol_lower}{Numeric value indicating the lower threshold for tolerance METs. Default is 1.25.}

\item{tol_upper}{Numeric value indicating the upper threshold for tolerance METs. Default is 3.}

\item{epoch}{Numeric value indicating the epoch length in minutes. Default is 1 minute.}
}
\value{
Integer vector indicating bouts of physical activity (1 for bout, 0 for non-bout).
}
\description{
This function identifies bouts of physical activity behaviors based on metabolic equivalent tasks (METs) data.
It accounts for periods of wear time and applies thresholds to determine valid bouts.
This is also a simpler R version of bout function from accelerometry package.
}
\examples{
mets <- c(1.5, 2.0, 1.0, 5.0, 6.0, 1.0, 2.5, 3.0, 1.0, 1.5)
wake <- c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1)
bouts(mets, wake, bout_length = 10, thresh_lower = 1.25, thresh_upper = 6, tol = 0, epoch = 1)
}
