% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_ap.R
\name{breaks_ap}
\alias{breaks_ap}
\title{Estimate the number of sit to stand transitions}
\usage{
breaks_ap(posture)
}
\arguments{
\item{posture}{A numeric vector (in any EPOCH), where \code{0} indicates sitting,
\code{1} indicates standings, and \code{2} indicates stepping.}
}
\value{
A numeric value indicating the number of breaks from sitting
}
\description{
This function estimates the number of sit to stand transitions, where values
of \code{0} indicates sitting, \code{1} indicates standings, and \code{2} indicates
stepping.
}
\details{
This is essentially the same as \code{\link[activpalProcessing:breaks.AP]{activpalProcessing::breaks.AP()}};
however, if there are *\emph{NO} sit to stand transitations a value of \code{0} is
returned instead of \code{NA}.
}
\examples{
x <- sample(x = c(0, 1, 2), size = 100, replace = TRUE, prob = rep(1/3, 3))
breaks_ap(x)

}
