% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dir.R
\name{check_dir}
\alias{check_dir}
\title{Checking current working directory}
\usage{
check_dir(path = ".")
}
\arguments{
\item{path}{A character vector of full path names; the default corresponds to
the working directory, \code{getwd()}. Also, \code{path} sets the working directory
with \code{setwd(path)}.}
}
\value{
Error message is returned if current working directory does not have
the appropriate sub-directories else nothing is returned.
}
\description{
Checking that current working directory has the sub-directories:
\code{DATA_Iowa_activPAL}, \code{DATA_Pitt_activPAL}, and \code{DATA_WVU_activPAL},
which contains all raw sleep diary and ActivPAL data.
}
\examples{
# check_dir() # works if current directory has sub-directories
# check_dir(path = ".") # works if current directory has sub-directories
# check_dir(path = "../") # error by moving out of current directory

}
\seealso{
\code{getwd()}, \code{dir()}
}
