% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_1s_epoch.R
\name{create_1s_epoch}
\alias{create_1s_epoch}
\title{Create a 1 second EPOCH data set}
\usage{
create_1s_epoch(data, good_days = 1:9, remove_days = FALSE)
}
\arguments{
\item{data}{Either a \code{process.data.events} or \code{merge.data.events} data set to
be converted to a one second EPOCH data set}

\item{good_days}{An integer vector denoting the wear days that have valid
data. The default is \code{1:9}, as for the Pregnancy 24/7 study the wear period
last nine days.}

\item{remove_days}{A logical denoting if merge data set should be subsetted
by the \code{good_days}. Default is FALSE.}
}
\value{
A \code{one.epoch.data} object, which is a data.frame that has the same
variables as the \code{data} parameter plus \code{mets1}, \code{mets30}, \code{mets60}, and
\itemize{
\item \code{date} which is just date version of the \code{time} variable and
\item \code{ap.posture} which is a replicate of the \code{activity} variable to conform to
previous work.
}
}
\description{
Transforming the events data to a one second EPOCH data that contains
expanded metrics on physical activity.
}
\details{
The \code{cumulativesteps} and \code{methrs} variables are renamed to conform
to previous work done in this field.
\itemize{
\item \code{cumulativesteps} to \code{steps}
\item \code{methrs} to \code{met.hrs}
}

One second, thirty second, and sixty second variants of MET hours are created
and named accordingly.
\itemize{
\item \code{mets1} for one second
\item \code{mets30} for thirty second
\item \code{mets60} for sixty second
}
}
\examples{
\dontrun{
process_dat <- process_events(data = dat)
sec_by_sec <- create_1s_epoch(data = process_dat)
}

}
