% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dict.R
\name{create_dict}
\alias{create_dict}
\title{Creates a data dictionary and switches out special characters}
\usage{
create_dict(x, out = NULL)
}
\arguments{
\item{x}{A named object.}

\item{out}{Either a path to a file, a connection, or literal data (either a
single string or a raw vector) to export the data dictionary. The default is
\code{NULL}, which does not result in the creation of data dictionary.}
}
\value{
If the \code{out} parameter is \code{NULL} then only character vector with the
new variable names is returned.
}
\description{
Creates a data dictionary for a data set that originally had special
characters as variable names.
}
\details{
The \code{x} must be a named object as the main purpose of this function
is create a data dictionary for a data set that has special characters in the
variable names.

Every time this function is executed the named object \code{x} alters the names to
remove special character. When the \code{out} parameter is specified then a data
dictionary is created with the original and new variable names. The current
version of this function adjusts for special characters such as spaces, "-",
"/", "\%", and "#". The naming convention used follows the
\href{https://style.tidyverse.org}{Tidyverse} style guide.
}
\examples{
y <- 1:4
names(y) <- c("a/a", "b-b", "\%c", "d#")
names(y) <- create_dict(y)

}
