% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_min_ap.R
\name{lit_min_ap}
\alias{lit_min_ap}
\title{Estimate time in light intensity activity}
\usage{
lit_min_ap(mets, posture, epoch = 1)
}
\arguments{
\item{mets}{A numeric vector of MET values.}

\item{posture}{A numeric vector of values \code{0}-\code{2} indicating: \code{0} for
sitting, \code{1} for standing, and \code{2} for stepping.}

\item{epoch}{A numeric value indicating what epoch (in seconds) the data are
in (e.g. \code{60} = 1 minute epochs). The default value is \code{1} second.}
}
\value{
A numeric value indicating hours spent in light intensity activity.
}
\description{
This function estimates time spent in light intensity activity. Light
intensity activity is activity in the standing or stepping posture.
}
\details{
This function is an adaptation of
\code{\link[activpalProcessing:lit.min.AP]{activpalProcessing::lit.min.AP()}}, where the dependencies on certain MET
values to determine standing and stepping is excluded.
}
\examples{
set.seed(1997)
n <- 100
mets <- stats::rgamma(n = n, shape = 2)
posture <- rep(NA, n)
for (i in seq_along(posture)) {
  if (mets[i] < 1.5) {
    posture[i] <- sample(c(0, 1), 1, TRUE, c(0.75, 0.25))
  } else if (mets[i] >= 1.5 && mets[i] < 3) {
    posture[i] <- sample(c(0, 1, 2), 1, TRUE, c(0.25, 0.5, 0.25))
  } else {
    posture[i] <- sample(c(1, 2), 1, TRUE, c(0.25, 0.75))
  }
}
lit_min_ap(mets = mets, posture = posture)
}
