% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc.R
\name{loc}
\alias{loc}
\title{Determine the data location}
\usage{
loc(subject)
}
\arguments{
\item{subject}{A character vector denoting the Pregnancy 24/7 subject.}
}
\value{
A \code{character} value that is the directory path for the \code{subject} of
interest. If \code{subject} starts with a
\itemize{
\item \code{1} then the start of the directory path will be \code{./DATA_Iowa_activPAL},
\item \code{5} then the start of the directory path will be \code{./DATA_Pitt_activPAL}, or
\item \code{8} then the start of the directory path will be \code{./DATA_WVU_activPAL}.
}
}
\description{
A helper function for the \code{read} and \code{write} functions that returns the
directory path based on the \code{subject} ID.
}
\details{
This function is meant to work within the \code{\link[=read_events]{read_events()}} and
\code{\link[=write_events]{write_events()}} functions based on no errors being produced by
\code{\link[=check_dir]{check_dir()}}.
}
\examples{
loc("1001-AB")
loc("5001-AB")
loc("8001-AB")

}
