% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_events.R
\name{merge_events}
\alias{merge_events}
\title{Merge event data together}
\usage{
merge_events(
  data,
  add_events,
  start_time = NA,
  end_time = NA,
  off_times = NULL,
  good_days = 1:9,
  remove_days = FALSE
)
}
\arguments{
\item{data}{A \code{process.data.events} object produced by \code{\link[=process_events]{process_events()}}}

\item{add_events}{A \code{data.frame} of events to merge with \code{data} that has the
same variable names as \code{data}.}

\item{start_time}{A date-time (POSIXct) value that denotes first date and
time that should be recorded in merged data set. Default is \code{NA} to indicate
that the merge data set does not need to be subsetted.}

\item{end_time}{A date-time (POSIXct) value that denotes last date and time
that should be recorded in merged data set. Default is \code{NA} to indicate
that the merge data set does not need to be subsetted.}

\item{off_times}{A \code{list} containing two objects of equal length denoting the
beginning and ending times of when the monitors were removed. For the
Pregnancy 24/7 study, the \code{\link[=windows_monitor]{windows_monitor()}} function was used to find these
times. Default is \code{NULL}, which indicates that is no monitor off times.}

\item{good_days}{An integer vector denoting the wear days that have valid
data. The default is \code{1:9}, as for the Pregnancy 24/7 study the wear period
last nine days.}

\item{remove_days}{A logical denoting if merge data set should be subsetted
by the \code{good_days}. Default is FALSE.}
}
\value{
A \code{merge.data.events} object, which is a \code{data.frame} with all of the
variables in \code{data} plus the \code{event} variable.
}
\description{
Merge events data set produced by \code{\link[=process_events]{process_events()}} with any extra events
that needed to be included by an external source, such as the Actiwatch and
sleep diary.
}
\details{
After \code{data} and \code{add_events} have been merged together, all missing
values created by merging the two data set together are filled in by their
appropriate values. For the looping variables, usually, the last value is
carried forwarded. As for the activity variables, it depends on how the
original variable was created as for how the missing values are filled in.

If there are monitor off times, the looping variables are replaced with the
value of \code{99} to indicate a special type of missing values related to
monitor wear.

There are two more objectives of this function.
\enumerate{
\item Round the \code{time} variable to the nearest second with
\code{\link[lubridate:round_date]{lubridate::round_date()}} and
\item create a new variable called \code{event} that denotes the event in recorded.
}
}
\examples{
# See process_data() for an example of how this function is used.

}
\seealso{
\code{\link[lubridate:round_date]{lubridate::round_date()}}
}
