% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_min_ap.R
\name{mpa_min_ap}
\alias{mpa_min_ap}
\alias{vpa_min_ap}
\alias{mvpa_min_ap}
\title{Levels of physical activity}
\usage{
mpa_min_ap(mets, epoch = 1)

vpa_min_ap(mets, epoch = 1)

mvpa_min_ap(mets, epoch = 1)
}
\arguments{
\item{mets}{A numeric vectors of METs (metabolic equivalents).}

\item{epoch}{Length of interval, default is \code{1} second.}
}
\value{
Numeric vector denoting the number of minutes.
}
\description{
Determines how many moderate, vigrous, and moderate-vigrous physical activity
minutes occurred for a given interval.
}
\details{
These functions are similar to the
\code{\link[activpalProcessing:mvpa.min.AP]{activpalProcessing::mvpa.min.AP()}} function but the METs cutoff values used
to define the levels of activity have be adjusted.
\itemize{
\item METs between [3, 6) define moderate physical activity
\item METs greater than or equal to 6 define vigrous physical activity
\item METs greater than or equal to 3 define moderate-vigrous physical activity
}
}
\examples{
x <- runif(100, 1, 10)
mpa_min_ap(x)
vpa_min_ap(x)
mvp_min_ap(x)

}
