% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapfix.R
\name{overlapfix}
\alias{overlapfix}
\title{Resolve Overlaps Between Sleep and Intervention Time Windows}
\usage{
overlapfix(sleepwindow, interwindow)
}
\arguments{
\item{sleepwindow}{A list or data frame containing \code{start} and \code{end}
POSIXct vectors. These are the baseline sleep intervals.}

\item{interwindow}{A list or data frame containing \code{start} and \code{end}
POSIXct vectors. These are the work or monitor_off window intervals to be adjusted.}
}
\value{
A list containing the adjusted \code{start} and \code{end} POSIXct vectors
for the \code{interwindow}.
}
\description{
Adjusts intervention time windows (\code{interwindow}) to avoid overlaps with
sleep time windows (\code{sleepwindow}).
}
\details{
This function operates by combining all start and end points into a single
timeline and sorting them chronologically. It iterates through the timeline
and, upon finding abnormal intersections between sleep and other windows
to resolve the overlap by swapping event orders or
shifting times by 60 seconds. For instance, an intervention window fully
contained within a sleep window may be removed.
}
\examples{
# Create baseline sleep intervals
sleep_intervals <- data.frame(
  start = as.POSIXct(c("2025-07-25 22:10:00", "2025-07-26 01:30:00")),
  end = as.POSIXct(c("2025-07-25 22:50:00", "2025-07-26 02:00:00"))
)

# Create intervention intervals to be adjusted
intervention_intervals <- data.frame(
  start = as.POSIXct(c(
    # 1. A non-overlapping interval
    "2025-07-25 21:00:00",
    # 2. An interval fully contained within a sleep period (target for removal)
    "2025-07-25 22:20:00",
    # 3. An interval with an overlapping start point (target for adjustment)
    "2025-07-26 01:50:00"
  )),
  end = as.POSIXct(c(
    "2025-07-25 21:30:00",
    "2025-07-25 22:30:00",
    "2025-07-26 02:10:00"
  ))
)

# Run the function
adjusted_intervals <- overlapfix(sleep_intervals, intervention_intervals)

# Check the results
print(adjusted_intervals)

}
