% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1s_epoch.R
\name{plot_1s_epoch}
\alias{plot_1s_epoch}
\title{Graphing events}
\usage{
plot_1s_epoch(data, sleep_times = NULL, plot = FALSE)
}
\arguments{
\item{data}{A \code{one.epoch.data} object created by \code{\link[=create_1s_epoch]{create_1s_epoch()}}.}

\item{sleep_times}{A \code{data.frame} that contains information on sleep and nap
windows during the wear period. The data set must contain the following
variables and the default is \code{NULL} denoting sleep data does not need to be
included.
\itemize{
\item a date-time object named \code{sleep_start} denoting sleep and nap onset
\item a date-time object named \code{sleep_stop} denoting sleep and nap wake times
\item an integer vector named \code{day} denoting wear day
\item a character vector named \code{label} denoting the type of sleep
}}

\item{plot}{A logical denoting where the plot should appear.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object that may be displayed or saved later.
}
\description{
Graphing the activity and sleep during the wear period.
}
\details{
Activities are generally grouped by 300 second intervals; however,
if the day does not last more than 300 seconds then the activities are
grouped by 50 second intervals. The specific acitivies examined in the plot
are standing, walking, and laying.
}
\examples{
\dontrun{
process_dat <- process_events(data = dat)
sec_by_sec <- create_1s_epoch(data = process_dat)
g <- plot_1s_epoch(sec_by_sec)
}
}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
