% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_daily}
\alias{process_daily}
\title{Process summary Data (Daily)}
\usage{
process_daily(subject, visit)
}
\arguments{
\item{subject}{Character string representing the subject ID.}

\item{visit}{Character string representing the visit ID.}

\item{dir}{Character string representing the directory where the raw data files are located.}

\item{save}{Character string representing the directory where the processed data should be saved.}
}
\value{
None. The function updates the weekly summary CSV file with the processed data.
}
\description{
This function processes daily data for a given subject and visit, summarizing the data and updating the summary CSV file.
}
\examples{
process_daily(subject = "1001-AA", visit = 1)
}
