% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_data_1sec}
\alias{process_data_1sec}
\title{Process data for the Pregnancy 24/7 study}
\usage{
process_data_1sec(
  subject,
  visit,
  plotsave = TRUE,
  day1 = TRUE,
  day2 = TRUE,
  day3 = TRUE,
  day4 = TRUE,
  day5 = TRUE,
  day6 = TRUE,
  day7 = TRUE,
  day8 = TRUE,
  day9 = TRUE
)
}
\arguments{
\item{subject}{A character vector denoting the Pregnancy 24/7 subject.}

\item{plotsave}{Logical value indicating whether to save the plot as a PDF.
Default is \code{TRUE}.}

\item{day1}{Logical denoting this a valid day of wear for day 1 of the wear
period. Default is \code{TRUE}.}

\item{day2}{Logical denoting this a valid day of wear for day 2 of the wear
period. Default is \code{TRUE}.}

\item{day3}{Logical denoting this a valid day of wear for day 3 of the wear
period. Default is \code{TRUE}.}

\item{day4}{Logical denoting this a valid day of wear for day 4 of the wear
period. Default is \code{TRUE}.}

\item{day5}{Logical denoting this a valid day of wear for day 5 of the wear
period. Default is \code{TRUE}.}

\item{day6}{Logical denoting this a valid day of wear for day 6 of the wear
period. Default is \code{TRUE}.}

\item{day7}{Logical denoting this a valid day of wear for day 7 of the wear
period. Default is \code{TRUE}.}

\item{day8}{Logical denoting this a valid day of wear for day 8 of the wear
period. Default is \code{TRUE}.}

\item{day9}{Logical denoting this a valid day of wear for day 9 of the wear
period. Default is \code{TRUE}.}
}
\description{
Process the Actiwatch, sleep diary, and activPAL data for the Pregnancy 24/7
study into 1 second epoch file and screening plot.
}
\details{
This function is based on the specific needs of the Pregnancy 24/7
study. Thus, the decision to only include parameters \code{day1} through \code{day9}
to denote valid wear days was because the wear period for Pregnancy 24/7 only
lasted nine days. If added \code{day} parameters are needed, this function can be
adjusted for that change by either
\itemize{
\item adding a specific \code{day} parameter or
\item creating a new parameter that is a logical vector that is length of wear
period and removing the \code{day} parameters.
}

The \code{day} parameters only need to be specified when a day is invalid and to
make that specification have \code{day = FALSE}.

The entire creation of this function depends on the other functions in the
\code{pregnancy247} package. There is no \code{R} object returned by this function, but
files are written to their appropriate locations based on the \code{subject} and
\code{trimester} parameters.

Due to nature of the observational data, there is an inclusion of a second
interval of napping for each wear day. This is implemented through
\code{windows_nap()}.
}
\examples{
\dontrun{
# 0001-AB ####
## Trimester 1 ####
process_data_1sec(subject = "0001-AB", visit = 1)
 

}

}
