% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_events.R
\name{process_events}
\alias{process_events}
\title{Manipulate the events data recorded by activPAL}
\usage{
process_events(data)
}
\arguments{
\item{data}{A \code{data.events} object produced by \code{read_events()}.}
}
\value{
A \code{process.data.events} object, which is a \code{data.frame} containing
the variables discussed in the \strong{Details}.
}
\description{
Manipulate and clean the events data record by activPAL and imported by
\code{read_events()}.
}
\details{
This functions rearranges the variables produced by the activPAL
software and selects the variables related to:
\itemize{
\item \code{time},
\item \code{datacount},
\item \code{interval},
\item \code{activity},
\item \code{cummulativesteps}, and
\item \code{methrs}.
}

The naming of variables is based on the archived package named
\href{https://cran.r-project.org/src/contrib/Archive/activpalProcessing/}{\code{activpalProcessing}}
created by \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5469371/}{Kate Lynden}.

Beyond variable selection, this function specifically:
\itemize{
\item doubles the \code{cumulativesteps} as the activPAL device is only on one leg,
\item transforms the \code{time} variable to a date-time (POSIXct) numeric vector
based on the UTC time-zone, and
\item creates several new variables to indicate:
\itemize{
\item wear day based on sleep onset (\code{wear_day}),
\item sleep windows (\code{sleep_loop}),
\item nap windows (\code{nap_loop}),
\item wake windows (\code{wake_loop}), and
\item work windows (\code{work_loop})
}
}

that are all missing values to be filled in later by the appropriate data.
}
\examples{
\dontrun{
dat <- read_events("0001-AB", "3")
process_dat <- process_events(data = dat)
}

}
\seealso{
\code{\link[lubridate:with_tz]{lubridate::with_tz()}}
}
