% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_functions.R
\name{process_weekly}
\alias{process_weekly}
\title{Process Weekly Physical Activity and Sleep Data}
\usage{
process_weekly(subject, visit, overwrite = TRUE)
}
\arguments{
\item{subject}{A character string representing the subject identifier.}

\item{visit}{A character string representing the visit identifier.}
}
\value{
This function does not explicitly return a value but overwrites the selected weekly summary file.
}
\description{
This function processes weekly physical activity and sleep data for a given subject and visit.
}
\examples{
Example usage (interactive - requires user input for directory and file selection)
process_weekly(subject = "1001-AA", visit = 1)

}
