% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_check}
\alias{quality_check}
\title{Outlier screening for sleep or nap Actiwatch data}
\usage{
quality_check(subject, trimester, nap = FALSE, cutpoint = 30)
}
\arguments{
\item{subject}{A character vector denoting the Pregnancy 24/7 subject.}

\item{trimester}{A character or integer vector denoting trimester of
pregnancy.}

\item{nap}{Logical denoting if nap data is to be imported and screened.
Default is FALSE.}

\item{cutpoint}{A numeric value denoting the amount of time in minutes to
see if nap and sleep overlap. Default is \code{30} minutes.}
}
\description{
Importing, screening, and exporting sleep or nap Actiwatch data.
}
\details{
If nap data is being screened then sleep data is imported as well
for checking of overlapping nap and sleep windows. This function relies
soley on the other \code{msleep} functions. So, for further details, see the
documentation of the functions listed below.
}
\examples{
\dontrun{
quality_check(subject = "1000-AB", trimester = 1)
}
}
\seealso{
\code{\link[=read_msleep]{read_msleep()}}, \code{\link[=check_msleep]{check_msleep()}}, \code{\link[=write_msleep]{write_msleep()}},
\code{\link[=write_all_msleep]{write_all_msleep()}}
}
