% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_events.R
\name{read_events}
\alias{read_events}
\title{Imports activity events data}
\usage{
read_events(subject, trimester, file = NULL, ...)
}
\arguments{
\item{subject}{A character vector denoting the Pregnancy 24/7 subject.}

\item{trimester}{A character or integer vector denoting trimester of
pregnancy.}

\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector). The default value is \code{NULL}, which
uses the \code{subject} and \code{trimester} parameters to find the appropriate
filename.}

\item{...}{Any other parameters needed for \code{\link[utils:read.table]{utils::read.csv()}}.}
}
\value{
A \code{data.events} object is returned, which is a \code{data.frame} that only
includes the raw activPAL EventsEx data for the \code{subject} of interest during
the \code{trimester} of interest.
}
\description{
Imports the CSV (semi-colon separted value) activPAL EventsEX file
for the \code{subject} and \code{trimester} of interest that was exported with the
\href{https://www.palt.com}{activPAL} software.
}
\details{
When \code{file} is the default value of \code{NULL} the \code{subject} and
\code{trimester} parameters along with the current working directory are used to
find the file that ends in \code{EventsEx.csv} by using the \code{\link[=loc]{loc()}} function.
As the purpose of this function is to work in the background of the
\code{\link[=process_data]{process_data()}} function, a filename should only be specified when
\code{read_events()} is used outside of the \code{\link[=process_data]{process_data()}}, usually for
individual examination of problematic data.

By default, the column names are skipped as the activPAL software tends to
produce variable names that do NOT conform to the standard \code{R} naming
convention.
}
\examples{
\dontrun{
# numeric trimester
dat <- read_events("0001-AB", 3)
}

\dontrun{
# charactter trimester
dat <- read_events("0001-AB", "3")
}

}
\seealso{
\code{\link[utils:read.table]{utils::read.csv()}}, \code{\link[=loc]{loc()}}
}
