% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msleep.R
\name{read_msleep}
\alias{read_msleep}
\title{Import master sleep and nap data.}
\usage{
read_msleep(file, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).}

\item{...}{any other parameters needed for \code{readxl::read_xlsx()}.}
}
\value{
A \code{msleep} object, which is a \code{data.frame} containing the imported
master sleep or nap data.
}
\description{
Importing the raw Microsoft Excel (.xlsx) master sleep and nap data exported
from the ActivWATCH device.
}
\details{
Empty strings (\code{""}), missing values (\code{NA}), and not a number
(\code{NaN}) values exported by the ActivWATCH device are all considered as
missing values for this importing process.
}
\examples{
# rest <- read_msleep("msleep.xlsx")

}
\seealso{
\code{readxl::read_xlsx()}
}
