% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot_1s_epoch.R
\name{save_plot_1s_epoch}
\alias{save_plot_1s_epoch}
\title{Export 1 second EPOCH plot}
\usage{
save_plot_1s_epoch(x, subject, trimester, file = NULL, ...)
}
\arguments{
\item{x}{A \code{\link[graphics:plot.default]{graphics::plot()}} or \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object to be exported.}

\item{subject}{A character vector denoting the Pregnancy 24/7 subject.}

\item{trimester}{A character or integer vector denoting trimester of
pregnancy.}

\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector). The default value is \code{NULL}, which
uses the \code{subject} and \code{trimester} parameters to find the appropriate
filename.}

\item{...}{Any other parameters needed for \code{\link[grDevices:pdf]{grDevices::pdf()}}}
}
\description{
Save the 1 second EPOCH plot as a PDF document based on the \code{subject} or
\code{file} parameter.
}
\details{
When \code{file} is the default value of \code{NULL} the \code{subject} and
\code{trimester} parameters along with the current working directory are used to
write the files that ends with \verb{_grpahic.pdf} by using the \code{\link[=loc]{loc()}} function.
As the purpose of this function is to work in the background of the
\code{\link[=process_data]{process_data()}} function, a filename should only be specified when
\code{save_plot_1s_epoch} is used outside of the \code{\link[=process_data]{process_data()}}, usually for
individual examination of problematic data.
}
\examples{
\dontrun{
process_dat <- process_events(data = dat)
sec_by_sec <- create_1s_epoch(data = process_dat)
g <- plot_1s_epoch(sec_by_sec)
save_plot_1s_epoch(x = g, subject = "0001-AB", 1, width = 11, height = 8.5)
}

}
\seealso{
\code{\link[=plot_1s_epoch]{plot_1s_epoch()}}, \code{\link[grDevices:pdf]{grDevices::pdf()}},
\code{\link[graphics:plot.default]{graphics::plot()}}
}
