% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_daily.R
\name{summarize_daily}
\alias{summarize_daily}
\title{Summarize daily activity}
\usage{
summarize_daily(
  subject,
  trimester,
  data_events,
  data_1s_epoch,
  all_days = 1:9,
  good_days = 1:9
)
}
\arguments{
\item{subject}{A character vector denoting the Pregnancy 24/7 subject.}

\item{trimester}{A character or integer vector denoting trimester of
pregnancy.}

\item{data_events}{Either a \code{process.data.events} or \code{merge.data.events} data
set.}

\item{data_1s_epoch}{A \code{one.epoch.data} object created by
\code{\link[=create_1s_epoch]{create_1s_epoch()}}.}

\item{all_days}{An integer vector denoting the all the possible wear days.
The default is 1:9, as for the Pregnancy 24/7 study the wear period last nine
days.}

\item{good_days}{An integer vector denoting the wear days that have valid
data. The default is \code{1:9}, as for the Pregnancy 24/7 study the wear period
last nine days.}
}
\value{
A named \code{list} with objects:
\itemize{
\item \code{daily} a \code{data.frame} containing daily summary of activities for all the
possible wear days
\item \code{avgs} a \code{data.frame} containing the averages of the activities during the
wear period.
}
}
\description{
The activities during the wear period are summarized using events and 1
second EPOCH data for the \code{subject} of interest during the \code{trimester} of
interest.
}
\details{
See the vingettes for a full description of the variables created.
}
\examples{
\dontrun{
process_dat <- process_events(data = dat)
sec_by_sec <- create_1s_epoch(data = process_dat)
summ_daily <- summarize_daily(
 subject = "0001-AB", trimester = "1",
 data_events = process_dat, data_1s_epoch = sec_by_sec
)
}

}
