% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\alias{write_events}
\alias{write_1s_epoch}
\alias{write_daily}
\alias{write_avgs}
\alias{write_all_avgs}
\title{Export processed events data}
\usage{
write_events(x, subject, trimester, file = NULL, ...)

write_1s_epoch(x, subject, trimester, file = NULL, ...)

write_daily(x, subject, trimester, file = NULL, ...)

write_avgs(x, subject, trimester, file = NULL, ...)

write_all_avgs(x, subject, trimester, file = NULL, ...)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{subject}{A character vector denoting the Pregnancy 24/7 subject.}

\item{trimester}{A character or integer vector denoting trimester of
pregnancy.}

\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector). The default value is \code{NULL}, which
uses the \code{subject} and \code{trimester} parameters to find the appropriate
filename.}

\item{...}{Any other parameters needed for \code{\link[readr:write_delim]{readr::write_csv()}}.}
}
\description{
Export events, 1 second EPOCH, daily summary, and wear period averages data
sets to \code{subject} specific folders.
}
\details{
When \code{file} is the default value of \code{NULL} the \code{subject} and
\code{trimester} parameters along with the current working directory are used to
write the files that ends with
\itemize{
\item \verb{_eventsfile.csv} for events data to be exported with \code{write_events},
\item \verb{_1secepoch.csv} for 1 second EPOCH data to be exported with
\code{write_1s_epoch},
\item \verb{_daily_values.csv} for the daily wear data to be exported with
\code{write_daily}, and
\item \verb{_weekly_avgs.csv} for the averages data to be exported with \code{write_avgs}
}

by using the \code{\link[=loc]{loc()}} function. As the purpose of these functions are to work
in the background of the \code{\link[=process_data]{process_data()}} function, a filename should only be
specified when \code{write_(datatype)} is used outside of the \code{\link[=process_data]{process_data()}},
usually for individual examination of problematic data.

In this collection of \code{write_(datatype)} functions, the \code{write_all_avgs}
function performs more tasks than simply writing a data set. This function
looks for data sets that already exists within the project directory that
contains a collection of wear period averages based on the site given by the
\code{subject} parameter. For Pregnancy 24/7, the data sets file names are
\itemize{
\item \code{weekly_avgs_IOWA.csv} for The University of Iowa,
\item \code{weekly_avgs_PITT.csv} for University of Pittsburgh, and
\item \code{weekly_avgs_WVU.csv} for West Virginia University.
}

If these files already exists, then the wear period averages for the
\code{subject} of interest during the \code{trimester} of interest is merged with the
the existing data set based on the site the \code{subject} visited.
}
\examples{
# See process_data() for an example of how these functions are used.

}
\seealso{
\code{\link[readr:write_delim]{readr::write_csv()}}
}
