% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_msleep.R
\name{check_msleep}
\alias{check_msleep}
\title{Screening the master sleep or nap data.}
\usage{
check_msleep(x, nap = FALSE)
}
\arguments{
\item{x}{A \code{msleep} object imported with \code{read_msleep}.}

\item{nap}{Logical denoting if master nap data is to be screened for outlying
values. Default is \code{FALSE}.}
}
\value{
A \code{msleep} object, which is a \code{data.frame} containing two additional
variables \code{flag1} and \code{flag2}.
}
\description{
Screening the master sleep or nap data imported with \code{read_msleep()} for
outlier values based on the expertise of
\href{https://www.education.pitt.edu/people/cek51}{Christopher E. Kline}.
}
